doc: added rules.md to write checks rules
The existing rules are described as an overview inside README.md.
The team needs a dedicated Markdown page to consolidate all the
check hits details and examples. This would help customers to easily
identify the hits and apply the amendment.

As described by Simon Sawert, current project phase focuses on
debugging and further develop wsl linter's program itself. Hence,
the documentations can change abruptly so we should keep it as
simple as a single Markdown file for now.

This patch adds the rules.md into `doc` directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>