Only force when to add (not remove) case trailing whitespace
This commit loosens up quite alot on how case blocks should look. The main
reason is because they're quite complex, especially if you consider
comments. A comment in the end of a case might (depending on ths Pos())
belong to the next case for example. Even gofmt handles indentation
for comments in case blocks in multiple ways.

From now on the end user may or may not end cases with newlines or
comments. With this commit a user can configure when you *must* end
with a whitespace to improve readability which essentially is what this
litner is about