Make error checking more laxed
Only perform error checking when there's a single comparison with a
binary operator comparing an error with `nil`. Both `!=` and `==` are
candidates for cuddling.

These should be cuddled if `err` exist on the line above

```go
if err != nil {}

if err == nil {}
```

Nothing else should trigger this lint, all of these should be left

```go
// More than error checking
if err != nil || true {}

// Technically an error checking, but we only check for binary
// operators, not expressions.
if errors.Is(err, SomeErr) {}

// More than one error checking
if err != nil || err2 != nil {}
```

This also matches [`gofumpt`](https://github.com/mvdan/gofumpt) which
has this rule:

> No empty lines before a simple error check
>
> ```go
> foo, err := processFoo()
>
> if err != nil {
> 	return err
> }
> ```
>
> ```go
> foo, err := processFoo()
> if err != nil {
> 	return err
> }
> ```