Ignore idents for constants, types, packages and nil
There are a lot of nodes that holds an `*ast.Ident` that's not a name
that's declared by the user in the code. This can be things like
constants such as `true` or `false`, or just a type name such as
`string` in the statement `var a string`.

We never want to treat them as idents when looking for intersection
because they're not related to the code written by the user.

In this example:

```go
var z string
defer func(x string) {}("")
```

There are no intersections, only `z` on 1st line and `x` on 2nd line,
but faulty identifying `string` on both lines result in a false negative
where a non existing intersection is found.

We now make use of the types info we get from the analysis pass and
ignore type names, constants, nil and package names.

Fixes #199