Add support for `newline-after-block`
This adds a new check type to enforce newline after any block (closing
`}`). In addition, it also makes the check for empty line while using
`case-max-lines` a bit more strict by now _not_ counting a trailing
comment as an empty line.

While doing this, there's a new column based logic to figure out where
the newline should go in the end of a case. If it's aligned with the
last statement of the case body the line is inserted after. If it's
aligned with then next case arm, the line is inserted before the
comment.

While working on this we're also moving to a line based check that uses
`LineStart` to make replacements and with that simplifying some comment
parsing since we only need to know the pos where the line starts. This
also allowed to add support to fix removal of whitespaces both before
_and_ after leading comments in blocks.